[FIX] fixing communication between terminals (sending coord)

+++ src/signal.c
@@ -5,11 +5,12 @@
 ** signal
 */
 
-#include <stdio.h>
 #include <signal.h>
 #include <stdlib.h>
 #include <sys/types.h>
 #include <unistd.h>
+#include "navy.h"
+#include "my.h"
 
 static void handler(int sig, siginfo_t *info)
 {
@@ -20,7 +21,7 @@ static void handler(int sig, siginfo_t *info)
     kill(info->si_pid, SIGUSR1);
 }
 
-static void handler_return(int sig, siginfo_t *info)
+static void handler_return(int sig, void *info, void *ucontext)
 {
     if (sig == SIGUSR1)
         recep_sig += 1;
@@ -31,6 +32,7 @@ int get_signal(void)
     struct sigaction act;
     int i = 0;
     char *bin = malloc(sizeof(char)*5);
+
     act.sa_flags = SA_SIGINFO;
     act.sa_sigaction = &handler;
     sigaction(SIGUSR1, &act, NULL);
@@ -39,7 +41,7 @@ int get_signal(void)
         pause();
         bin[i] = recep_sig + '0';
         i++;
-    }
+        }
     bin[4] = 0;
     bin = clean_bin(bin);
     return (my_btoi(my_revstr(bin)));
@@ -47,16 +49,21 @@ int get_signal(void)
 
 void my_send_sig(int data, int pid)
 {
+    static int error = -1;
     struct sigaction act;
+
     act.sa_flags = SA_SIGINFO;
     act.sa_sigaction = &handler_return;
     recep_sig = 0;
     sigaction(SIGUSR1, &act, NULL);
     for (; recep_sig <= 3;) {
+        if (error == recep_sig)
+            break;
+        error = recep_sig;
         if ((data >> recep_sig) % 2 == 1)
             kill(pid, SIGUSR1);
         if ((data >> recep_sig) % 2 == 0)
             kill(pid, SIGUSR2);
-        usleep(500);
+        usleep(20000);
     }
 }
